<refact> : profileAPI 데이터 요청 코드 파일 분리
- 유저 profile과 관련한 데이터를 받아오는 API 코드를 분리하여 관리하고자 Profile 폴더를 생성및 분리하였음.

- issue #195
