+++ src/api/profileAPI.js
@@ -1,2 +1,16 @@
-// import BASE_URL from '../utills/baseUrl';
-// const profileAPI = {};
+import BASE_URL from '../utils/baseUrl';
+
+const profileAPI = {
+  async getUserInfo(token, accountname) {
+    const response = await fetch(`${BASE_URL}/profile/${accountname}`, {
+      method: 'GET',
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-type': 'application/json',
+      },
+    });
+    const data = await response.json();
+    return data;
+  },
+};
+export default profileAPI;
